+++ toolium/jira.py
@@ -17,16 +17,30 @@ limitations under the License.
 """
 
 import logging
+import re
 
 import requests
 
 from toolium.config_driver import get_error_message_from_exception
 from toolium.driver_wrappers_pool import DriverWrappersPool
 
-"""Configuration"""
-logger = logging.getLogger(__name__)
-# Dict to save tuples with jira keys, their test status and comments
+# Dict to save tuples with jira keys, their test status, comments and attachments
 jira_tests_status = {}
+
+# List to save temporary test attachments
+attachments = []
+
+# Jira configuration
+enabled = None
+execution_url = None
+summary_prefix = None
+labels = None
+comments = None
+fix_version = None
+build = None
+only_if_changes = None
+
+logger = logging.getLogger(__name__)
 
 
 def jira(test_key):
@@ -38,6 +52,7 @@ def jira(test_key):
 
     def decorator(test_item):
         def modified_test(*args, **kwargs):
+            save_jira_conf()
             try:
                 test_item(*args, **kwargs)
             except Exception as e:
@@ -53,6 +68,30 @@ def jira(test_key):
     return decorator
 
 
+def save_jira_conf():
+    """Read Jira configuration from properties file and save it"""
+    global enabled, execution_url, summary_prefix, labels, comments, fix_version, build, only_if_changes, attachments
+    config = DriverWrappersPool.get_default_wrapper().config
+    enabled = config.getboolean_optional('Jira', 'enabled')
+    execution_url = config.get_optional('Jira', 'execution_url')
+    summary_prefix = config.get_optional('Jira', 'summary_prefix')
+    labels = config.get_optional('Jira', 'labels')
+    comments = config.get_optional('Jira', 'comments')
+    fix_version = config.get_optional('Jira', 'fixversion')
+    build = config.get_optional('Jira', 'build')
+    only_if_changes = config.getboolean_optional('Jira', 'onlyifchanges')
+    attachments = []
+
+
+def add_attachment(attachment):
+    """ Add a file path to attachments list
+
+    :param attachment: attachment file path
+    """
+    if attachment:
+        attachments.append(attachment)
+
+
 def add_jira_status(test_key, test_status, test_comment):
     """Save test status and comments to update Jira later
 
@@ -60,78 +99,82 @@ def add_jira_status(test_key, test_status, test_comment):
     :param test_status: test case status
     :param test_comment: test case comments
     """
-    if test_status == 'Fail':
-        if test_key in jira_tests_status and jira_tests_status[test_key][2]:
-            test_comment = '{}\n{}'.format(jira_tests_status[test_key][2], test_comment)
-        jira_tests_status[test_key] = (test_key, 'Fail', test_comment)
-    elif test_status == 'Pass':
-        # Don't overwrite previous fails
-        if test_key not in jira_tests_status:
-            jira_tests_status[test_key] = (test_key, 'Pass', test_comment)
+    global attachments
+    if test_key and enabled:
+        if test_key in jira_tests_status:
+            # Merge data with previous test status
+            previous_status = jira_tests_status[test_key]
+            test_status = 'Pass' if previous_status[1] == 'Pass' and test_status == 'Pass' else 'Fail'
+            if previous_status[2] and test_comment:
+                test_comment = '{}\n{}'.format(previous_status[2], test_comment)
+            elif previous_status[2] and not test_comment:
+                test_comment = previous_status[2]
+            attachments += previous_status[3]
+        # Add or update test status
+        jira_tests_status[test_key] = (test_key, test_status, test_comment, attachments)
 
 
 def change_all_jira_status():
     """Iterate over all jira test cases, update their status in Jira and clear the dictionary"""
     for test_status in jira_tests_status.values():
-        change_jira_status_with_config(*test_status)
+        change_jira_status(*test_status)
     jira_tests_status.clear()
 
 
-def change_jira_status_with_config(test_key, test_status, test_comment):
-    """Read Jira configuration properties and update test status in Jira
+def change_jira_status(test_key, test_status, test_comment, test_attachments):
+    """Update test status in Jira
 
     :param test_key: test case key in Jira
     :param test_status: test case status
     :param test_comment: test case comments
+    :param test_attachments: test case attachments
     """
-    config = DriverWrappersPool.get_default_wrapper().config
-    if config.getboolean_optional('Jira', 'enabled'):
-        execution_url = config.get_optional('Jira', 'execution_url')
-        summary_prefix = config.get_optional('Jira', 'summary_prefix')
-        labels = config.get_optional('Jira', 'labels')
-        comments = config.get_optional('Jira', 'comments')
-        if test_comment:
-            comments = '{}\n{}'.format(comments, test_comment) if comments else test_comment
-        fix_version = config.get_optional('Jira', 'fixversion')
-        build = config.get_optional('Jira', 'build')
-        only_if_changes = config.getboolean_optional('Jira', 'onlyifchanges')
-        change_jira_status(execution_url, test_key, test_status, summary_prefix, labels, comments, fix_version, build,
-                           only_if_changes)
+    if not execution_url:
+        logger.warn("Test Case '{}' can not be updated: execution_url is not configured".format(test_key))
+        return
 
-
-def change_jira_status(execution_url, test_key, test_status, summary_prefix=None, labels=None, comments=None,
-                       fix_version=None, build=None, only_if_changes=False):
-    """Update test status in Jira
-
-    :param execution_url: url of the jira test case execution service
-    :param test_key: test case key in Jira
-    :param test_status: test case status
-    :param summary_prefix: test case summary prefix
-    :param labels: test case labels
-    :param comments: test case comments
-    :param fix_version: test case fix version
-    :param build: test case build
-    :param only_if_changes:
-        if true, only create a new execution if the test status has changed
-        if false, create a new execution always
-    """
     logger.info("Updating Test Case '{0}' in Jira with status {1}".format(test_key, test_status))
-    payload = {'jiraTestCaseId': test_key, 'jiraStatus': test_status, 'summaryPrefix': summary_prefix, 'labels': labels,
-               'comments': comments,
-               'version': fix_version, 'build': build}
+    composed_comments = comments
+    if test_comment:
+        composed_comments = '{}\n{}'.format(comments, test_comment) if comments else test_comment
+    payload = {'jiraTestCaseId': test_key, 'jiraStatus': test_status, 'summaryPrefix': summary_prefix,
+               'labels': labels, 'comments': composed_comments, 'version': fix_version, 'build': build}
     if only_if_changes:
         payload['onlyIfStatusChanges'] = 'true'
     try:
-        response = requests.get(execution_url, params=payload)
-    except Exception as e:
-        if execution_url:
-            logger.warn("Error updating Test Case '{}' using execution_url '{}': {}".format(test_key, execution_url, e))
+        if test_attachments and len(test_attachments) > 0:
+            files = dict()
+            for index in range(len(test_attachments)):
+                files['attachments{}'.format(index)] = open(test_attachments[index], 'rb')
         else:
-            logger.warn("Error updating Test Case '{}': execution_url is not configured".format(test_key))
+            files = None
+        response = requests.post(execution_url, data=payload, files=files)
+    except Exception as e:
+        logger.warn("Error updating Test Case '{}': {}".format(test_key, e))
         return
 
-    logger.debug("Request url: {}".format(response.url))
     if response.status_code >= 400:
-        logger.warn("Error updating Test Case '{}': [{}] {}".format(test_key, response.status_code, response.content))
+        logger.warn("Error updating Test Case '{}': [{}] {}".format(test_key, response.status_code,
+                                                                    get_error_message(response.content)))
     else:
-        logger.debug("Response content: {}".format(response.content.decode().splitlines()[0]))
+        logger.debug("{}".format(response.content.decode().splitlines()[0]))
+
+
+def get_error_message(response_content):
+    """Extract error message from the HTTP response
+
+    :param response_content: HTTP response from test case execution API
+    :returns: error message
+    """
+    apache_regex = re.compile('.*<u>(.*)</u></p><p>.*')
+    match = apache_regex.search(response_content)
+    if match:
+        error_message = match.group(1)
+    else:
+        local_regex = re.compile('.*<title>(.*)</title>.*')
+        match = local_regex.search(response_content)
+        if match:
+            error_message = match.group(1)
+        else:
+            error_message = response_content
+    return error_message