+++ toolium/test/test_jira.py
@@ -16,6 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+import os
 import unittest
 
 import mock
@@ -24,32 +25,97 @@ from nose.tools import assert_equal, assert_in, assert_raises
 from requests.exceptions import ConnectionError
 
 from toolium import jira
+from toolium.driver_wrappers_pool import DriverWrappersPool
 
 
 class JiraTests(unittest.TestCase):
+    def tearDown(self):
+        # Clear jira module configuration
+        jira.enabled = None
+        jira.execution_url = None
+        jira.summary_prefix = None
+        jira.labels = None
+        jira.comments = None
+        jira.fix_version = None
+        jira.build = None
+        jira.only_if_changes = None
+        jira.jira_tests_status.clear()
+        jira.attachments = []
+
     @mock.patch('toolium.jira.logger')
     @requests_mock.Mocker()
     def test_change_jira_status(self, jira_logger, req_mock):
-        # Test input
-        execution_url = 'http://server/execution_service'
+        # Test response
         response = b"The Test Case Execution 'TOOLIUM-2' has been created\r\n"
 
+        # Configure jira module
+        jira.enabled = True
+        jira.execution_url = 'http://server/execution_service'
+        jira.summary_prefix = 'prefix'
+        jira.labels = 'label1 label2'
+        jira.comments = 'comment'
+        jira.fix_version = 'Release 1.0'
+        jira.build = '453'
+        jira.only_if_changes = True
+
         # Configure mock
-        req_mock.get(execution_url, content=response)
+        req_mock.post(jira.execution_url, content=response)
 
         # Test method
-        jira.change_jira_status(execution_url, test_key='TOOLIUM-1', test_status='Pass', summary_prefix='prefix',
-                                labels='label1 label2', comments='comment',
-                                fix_version='Release 1.0', build='453', only_if_changes=True)
+        jira.change_jira_status('TOOLIUM-1', 'Pass', None, [])
 
         # Check requested url
-        for partial_url in [execution_url, 'jiraStatus=Pass', 'jiraTestCaseId=TOOLIUM-1', 'summaryPrefix=prefix',
+        assert_equal(jira.execution_url, req_mock.request_history[0].url)
+        for partial_url in ['jiraStatus=Pass', 'jiraTestCaseId=TOOLIUM-1', 'summaryPrefix=prefix',
                             'labels=label1+label2', 'comments=comment', 'version=Release+1.0', 'build=453',
                             'onlyIfStatusChanges=true']:
-            assert_in(partial_url, req_mock.request_history[0].url)
+            assert_in(partial_url, req_mock.request_history[0].text)
 
         # Check that binary response has been decoded
-        expected_response = "Response content: The Test Case Execution 'TOOLIUM-2' has been created"
+        expected_response = "The Test Case Execution 'TOOLIUM-2' has been created"
+        jira_logger.debug.assert_called_with(expected_response)
+
+    @mock.patch('toolium.jira.logger')
+    @requests_mock.Mocker()
+    def test_change_jira_status_attachments(self, jira_logger, req_mock):
+        # Test response
+        response = b"The Test Case Execution 'TOOLIUM-2' has been created\r\n"
+
+        # Configure jira module
+        jira.enabled = True
+        jira.execution_url = 'http://server/execution_service'
+        jira.summary_prefix = 'prefix'
+        jira.labels = 'label1 label2'
+        jira.comments = 'comment'
+        jira.fix_version = 'Release 1.0'
+        jira.build = '453'
+        jira.only_if_changes = True
+        resources_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'resources')
+        attachments = [os.path.join(resources_path, 'ios.png'), os.path.join(resources_path, 'ios_web.png')]
+
+        # Configure mock
+        req_mock.post(jira.execution_url, content=response)
+
+        # Test method
+        jira.change_jira_status('TOOLIUM-1', 'Pass', None, attachments)
+
+        # Check requested url
+        body_bytes = req_mock.last_request.body
+        try:
+            # Python 3
+            body = "".join(map(chr, body_bytes))
+        except TypeError:
+            # Python 2.7
+            body = body_bytes
+        for partial_url in ['"jiraStatus"\r\n\r\nPass', '"jiraTestCaseId"\r\n\r\nTOOLIUM-1',
+                            '"summaryPrefix"\r\n\r\nprefix', '"labels"\r\n\r\nlabel1 label2',
+                            '"comments"\r\n\r\ncomment', '"version"\r\n\r\nRelease 1.0', '"build"\r\n\r\n453',
+                            '"onlyIfStatusChanges"\r\n\r\ntrue', '"attachments0"; filename="ios.png"',
+                            '"attachments1"; filename="ios_web.png"']:
+            assert_in(partial_url, body)
+
+        # Check that binary response has been decoded
+        expected_response = "The Test Case Execution 'TOOLIUM-2' has been created"
         jira_logger.debug.assert_called_with(expected_response)
 
     @mock.patch('toolium.jira.logger')
@@ -58,56 +124,99 @@ class JiraTests(unittest.TestCase):
         # Configure jira mock
         jira_get.side_effect = ConnectionError('exception error')
 
+        # Configure jira module
+        jira.enabled = True
+
         # Test method
-        jira.change_jira_status('', 'TOOLIUM-1', 'Pass')
+        jira.change_jira_status('TOOLIUM-1', 'Pass', None, [])
 
         # Check logging error message
-        expected_response = "Error updating Test Case 'TOOLIUM-1': execution_url is not configured"
+        expected_response = "Test Case 'TOOLIUM-1' can not be updated: execution_url is not configured"
         jira_logger.warn.assert_called_with(expected_response)
 
     @mock.patch('toolium.jira.logger')
-    @mock.patch('toolium.jira.requests.get')
-    def test_change_jira_status_exception(self, jira_get, jira_logger):
+    @mock.patch('toolium.jira.requests.post')
+    def test_change_jira_status_exception(self, jira_post, jira_logger):
         # Configure jira mock
-        jira_get.side_effect = ConnectionError('exception error')
+        jira_post.side_effect = ConnectionError('exception error')
+
+        # Configure jira module
+        jira.enabled = True
+        jira.execution_url = 'http://server/execution_service'
 
         # Test method
-        jira.change_jira_status('http://server/execution_service', 'TOOLIUM-1', 'Pass')
+        jira.change_jira_status('TOOLIUM-1', 'Pass', None, [])
 
         # Check logging error message
-        expected_response = "Error updating Test Case 'TOOLIUM-1' using execution_url " \
-                            "'http://server/execution_service': exception error"
+        expected_response = "Error updating Test Case 'TOOLIUM-1': exception error"
         jira_logger.warn.assert_called_with(expected_response)
 
     def test_jira_annotation_pass(self):
+        # Configure jira module
+        config = DriverWrappersPool.get_default_wrapper().config
+        try:
+            config.add_section('Jira')
+        except Exception:
+            pass
+        config.set('Jira', 'enabled', 'true')
+
         # Execute method with jira annotation
-        jira.jira_tests_status.clear()
         MockTestClass().mock_test_pass()
 
         # Check jira status
-        expected_status = {'TOOLIUM-1': ('TOOLIUM-1', 'Pass', None)}
+        expected_status = {'TOOLIUM-1': ('TOOLIUM-1', 'Pass', None, [])}
         assert_equal(expected_status, jira.jira_tests_status)
 
     def test_jira_annotation_fail(self):
+        # Configure jira module
+        config = DriverWrappersPool.get_default_wrapper().config
+        try:
+            config.add_section('Jira')
+        except Exception:
+            pass
+        config.set('Jira', 'enabled', 'true')
+
         # Execute method with jira annotation
-        jira.jira_tests_status.clear()
         assert_raises(AssertionError, MockTestClass().mock_test_fail)
 
         # Check jira status
-        expected_status = {'TOOLIUM-3': ('TOOLIUM-3', 'Fail', "The test 'test name' has failed: test error")}
+        expected_status = {'TOOLIUM-3': ('TOOLIUM-3', 'Fail', "The test 'test name' has failed: test error", [])}
         assert_equal(expected_status, jira.jira_tests_status)
 
     def test_jira_annotation_multiple(self):
+        # Configure jira module
+        config = DriverWrappersPool.get_default_wrapper().config
+        try:
+            config.add_section('Jira')
+        except Exception:
+            pass
+        config.set('Jira', 'enabled', 'true')
+
         # Execute methods with jira annotation
-        jira.jira_tests_status.clear()
         MockTestClass().mock_test_pass()
         assert_raises(AssertionError, MockTestClass().mock_test_fail)
         MockTestClass().mock_test_pass()
 
         # Check jira status
-        expected_status = {'TOOLIUM-1': ('TOOLIUM-1', 'Pass', None),
-                           'TOOLIUM-3': ('TOOLIUM-3', 'Fail', "The test 'test name' has failed: test error"),
-                           'TOOLIUM-1': ('TOOLIUM-1', 'Pass', None)}
+        expected_status = {'TOOLIUM-1': ('TOOLIUM-1', 'Pass', None, []),
+                           'TOOLIUM-3': ('TOOLIUM-3', 'Fail', "The test 'test name' has failed: test error", []),
+                           'TOOLIUM-1': ('TOOLIUM-1', 'Pass', None, [])}
+        assert_equal(expected_status, jira.jira_tests_status)
+
+    def test_jira_disabled(self):
+        # Configure jira module
+        config = DriverWrappersPool.get_default_wrapper().config
+        try:
+            config.add_section('Jira')
+        except Exception:
+            pass
+        config.set('Jira', 'enabled', 'false')
+
+        # Execute method with jira annotation
+        MockTestClass().mock_test_pass()
+
+        # Check jira status
+        expected_status = {}
         assert_equal(expected_status, jira.jira_tests_status)